refactor add note functionality: update note model to use double for color, enhance loading state handling in UI components

+++ lib/models/note_model.dart
@@ -12,7 +12,7 @@ class NoteModel {
   @HiveField(2)
   final String date;
   @HiveField(3)
-  final int color;
+  final double color;
 
   NoteModel(
       {required this.title,
